Working on DataProvider

+++ SolidCP/Sources/SolidCP.EnterpriseServer.Code/ExchangeServer/ExchangeServerController.cs
@@ -1456,7 +1456,7 @@ namespace SolidCP.EnterpriseServer
             string accountName, string displayName, string primaryEmailAddress, bool mailEnabledPublicFolder,
             MailboxManagerActions mailboxManagerActions, string samAccountName, string accountPassword, int mailboxPlanId, string subscriberNumber)
         {
-            return Database.AddExchangeAccount(itemId, (int)accountType,
+            return Database.AddExchangeAccount(itemId, accountType,
                 accountName, displayName, primaryEmailAddress, mailEnabledPublicFolder,
                 mailboxManagerActions.ToString(), samAccountName, mailboxPlanId, (string.IsNullOrEmpty(subscriberNumber) ? null : subscriberNumber.Trim()));
         }
